cover-letter: iommu: Pass in old_domain pointer to attach_dev
Some cleanups related to release_domain, and reduce the amount of callers
of iommu_get_domain_for_dev() in the drivers.

One goal is to keep the group->mutex lock consistent at every attachment.
The arm_smmu_release_device() handling an attachment was unnecessary, and
cannot take the group->mutex as one of the callers invokes it even before
the iommu_group is created. So, using the release_domain could avoid this
awkward situation.

This is also the part-1 preparatory series derived from:
https://lore.kernel.org/all/cover.1756682135.git.nicolinc@nvidia.com/