cover-letter: iommu: Introduce and roll out test_dev domain op
Add a new test_dev domain op for drivers to run a compatibility test prior
to an actual attachement at the driver level. Any incompatible attachment
will be rejected early, allowing the iommu core to postpone any concurrent
attachement during a device reset state.

This is a preparatory series for new iommu_dev_reset APIs:
https://lore.kernel.org/all/cover.1756682135.git.nicolinc@nvidia.com/

This is on github:
https://github.com/nicolinc/iommufd/tree/iommu_test_dev