iommufd: HACK keep a iommu device pointer around
Currently attaching to a dmabuf cannot do anything except dma map to a
struct device and return a dma_addr_t. This is useless for iommufd, but we
can make something demonstrable by passing the struct device of an iommu
device.

iommu devices always cause the dma_api to make dma_addr_t == phys_addr_t,
and several of the drivers have assertions to this effect.

It will probably fail p2p checks though.

This hack is to be replaced with a negotiation of address format, such as
this proposal:

https://lore.kernel.org/all/20250926084624.2288-1-thomas.hellstrom@linux.intel.com/

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>