iommufd/device: Make iommufd_device_is_attached non-static
A new SET_OPTION will reuse this helper for a sanity check before setting
a per-idev property.

Given that the attach handle can be NULL if device is not attached, add a
pointer check prior to the xa_load(). This is not a problem currently, as
the only caller iommufd_device_do_replace() verifies the attach handle.

Also, add lockdep_assert_held on igroup's mutex;

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>