vfio/nvgrace: Support get_dmabuf_phys
Call vfio_pci_core_fill_phys_vec() with the proper physical ranges
for the synthetic BAR 2 and BAR 4 regions. Otherwise use the normal
flow based on the PCI bar.

FIXME: This creates a weirdness where the size of the DMABUF is smaller
than the size of the bar. I'm not sure what to do about this but it
does seem like the right thing to do. Userspace would need to discover
the smaller size somehow? Is there an ioctl on dmabuf already?

FIXME: provider = NULL as cachable probably needs some more fixing

NOTE: all the bar 2 and bar 4 memory actually works for all P2P, but
we don't have a way to express this in the p2p framework right now. Doesn't
matter for our use cases.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>