arm-smmu-v3: Store ASIDs and VMID in arm_smmu_master
Currently, ASID is allocated per smmu_domain, stored in the domain, and
freed along with the domain.

Practically, ASID is only used in a CD as an iotlb tag. Therefore, ASID
doesn't really follow the life cycle of a domain but domain attachment.

This applies to the VMID as well.

To reflect that, both ASID and VMID should be decoupled from the domain.

A device can only have one ASID per SSID and one VMID per SID. Thus, add
an ASID array and VMID in the arm_smmu_master structure.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>