iommu/arm-smmu-v3: Drop STRTAB_STE_1_MEV in arm_smmu_get_ste_used() for S2
The MEV bit was introduced for nesting case only, to mitigate a potential
user space DoS on eventq. And only nested S1 would set this bit. But it's
added to the S2's used list as well.

This would result in ste->word[1] becoming critical every time a device is
switching between S1+S2 mode (that always sets MEV) and S2-only mode (that
never sets MEV). So, it won't never do a hitless update.

However, for example, when a device running in VM attaches to an identity
domain but enables ATS for SVA cases, its physical STE should swtich from
S2-only mode to S1+S2 mode, going through the hitless path. But because of
the difference at the MEV field, it will invalidate the entire STE, which
breaks the ongoning DMA and raises C_BAD_STE fault.

Fix this by dropping it from the S2's list in arm_smmu_get_ste_used().

Fixes: da0c56520e88 ("iommu/arm-smmu-v3: Set MEV bit in nested STE for DoS mitigations")
Cc: stable@vger.kernel.org
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>