iommu/arm-smmu-v3: Always set EATS for nested S1 bypass domain
When a device attaches to an S2 domain direclty, the EATS is always set,
so long as the device supports ATS.

However, when a device attaches to a nested S1 bypass domain, though it
runs in the physical S2-only mode, EATS might not be set. However, when
it switches to a nested S1 translate domain (physical S1+S2 mode), EATS
might be set.

This would result in ste->word[1] becoming critical every time a device is
switching between S1+S2 mode (that might set EATS) and S2-only mode (that
doesn't set EATS). So, it won't never do a hitless update.

However, for example, when a device running in VM attaches to an identity
domain but enables ATS for SVA cases, its physical STE should swtich from
S2-only mode to S1+S2 mode, going through the hitless path. But because of
the difference at the EATS field, it will invalidate the entire STE, which
breaks the ongoning DMA and raises C_BAD_STE fault.

Fix this by always setting EATS in nesting S2 case, aligning with a normal
S2 case.

Fixes: 1e8be08d1c91 ("iommu/arm-smmu-v3: Support IOMMU_DOMAIN_NESTED")
Cc: stable@vger.kernel.org
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>